Extract JsonSchemaProvider from TemplateEngine (#7335)
* refactor template engine

* address comments
